requirements.txt
flask>=3.0.0
requests>=2.32.0
python-dotenv>=1.0.1
openai>=1.40.0
google-generativeai>=0.7.0
pydantic>=2.7.0
